Export plot/report as binaries
This allows us to invoke them in Ember Macro Benchmark more clearly as
these will be placed in ./node_modules/.bin/ on EMB installs which is in
$PATH.

Unfortunately, I couldn't find a good way to deal with ResultSets.R as
it is a binary dependencies but isn't an executable itself. And the
`source` in the plot.R and report.R that looks for ResultsSets.R isn't a
file but string thing.  idk

Seems good for now, maybe we can introduce a more robust build step
later that handles this.  To make the maintenance of these reports
easier on Kris.